Refactor Profile component and remove unused lists and layout components
